565 - client - Reorganize
